+++ src/layout/SigningActions/PanelAwaitingCurrentUserSignature.tsx
@@ -7,7 +7,7 @@ import { Button } from 'src/app-components/Button/Button';
 import { Spinner } from 'src/app-components/loading/Spinner/Spinner';
 import { Panel } from 'src/app-components/Panel/Panel';
 import { LiveValidationMessage } from 'src/app-components/ValidationMessage/LiveValidationMessage';
-import { useRequestFocus } from 'src/core/contexts/ElementFocus';
+import { useRequestFocus } from 'src/core/contexts/ElementFocusProvider';
 import { useIsAuthorized } from 'src/features/instance/useProcessQuery';
 import { UnknownError } from 'src/features/instantiate/containers/UnknownError';
 import { Lang } from 'src/features/language/Lang';
