+++ src/layout/SigningActions/PanelSigning.tsx
@@ -6,7 +6,7 @@ import { Dialog, Heading, Paragraph } from '@digdir/designsystemet-react';
 import { Button } from 'src/app-components/Button/Button';
 import { Panel } from 'src/app-components/Panel/Panel';
 import { LiveValidationMessage } from 'src/app-components/ValidationMessage/LiveValidationMessage';
-import { useFocusOnRequest } from 'src/core/contexts/ElementFocus';
+import { useFocusOnRequest } from 'src/core/contexts/ElementFocusProvider';
 import { useProcessNext } from 'src/features/instance/useProcessNext';
 import { useIsAuthorized } from 'src/features/instance/useProcessQuery';
 import { Lang } from 'src/features/language/Lang';
