+++ src/layout/SigningActions/SigningActionsComponent.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 import { FatalError } from 'src/app-components/error/FatalError/FatalError';
 import { Spinner } from 'src/app-components/loading/Spinner/Spinner';
 import { Panel } from 'src/app-components/Panel/Panel';
-import { ElementFocusProvider } from 'src/core/contexts/ElementFocus';
+import { ElementFocusProvider } from 'src/core/contexts/ElementFocusProvider';
 import { useIsAuthorized } from 'src/features/instance/useProcessQuery';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
